==> notebooks/1_data_wrangling.ipynb <==
# Data Wrangling

Step one is load up all the libraries I'll need for wrangling.

import pandas as pd
import numpy as np
import warnings
import re
import sagemaker
import boto3
import nltk
nltk.download('wordnet')
nltk.download('stopwords')
from nltk.tokenize import WhitespaceTokenizer
from nltk.stem.wordnet import WordNetLemmatizer
from nltk.corpus import stopwords
stop = stopwords.words('english')
from library.sb_utils import save_file
warnings.filterwarnings("ignore")

Next, I'll load the data. Right now, the data is split into reviews and products, and I need to consolidate them into a single dataframe. Also since both these files are local to my machine, I will also upload them an AWS S3 bucket.

reviews = pd.read_csv('../data/combined/reviews.csv')
products = pd.read_csv('../data/combined/products.csv')

sm_boto3 = boto3.client('sagemaker')
sess = sagemaker.Session()
region = sess.boto_session.region_name
bucket = 'ice-cream-reviews-sagemakerbucket'
print('Using bucket ' + bucket)

## Creating A Single DataFrame

So there are two different dataframes to work with. They are `reviews` and `products`. Unsurprisingly, these both correspond to the individual reviews and the details of each flavor. I'll start by looking at the first few rows of each dataframe.

reviews.head(3)

products.head(3)

So I can already see some overlapping information as well as informationthat is different in each dataframe. I also see that there are some columns that don't interest me for this project. So, to better prepare the data, I need to get all the columns I want from both dataframes and put it into a single dataframe. Before I create any new dataframes, I need to make sure that the the keys align.

set(products['key']) == set(reviews['key'])

Perfect! Now I can perform the merge. I'll store all the columns that I want into a dataframe called `data`.

data = pd.DataFrame()
text = []
title = []
brand = []
author = []
names = []
stars = []
ratings = []
rating_counts = []

for i in range(len(reviews)):
    text.append(reviews['text'].iloc[i])
    title.append(reviews['title'].iloc[i])
    brand.append(reviews['brand'].iloc[i])
    author.append(reviews['author'].iloc[i])
    stars.append(reviews['stars'].iloc[i])
    key = reviews['key'].iloc[i]
    products_key = products[products['key'] == key]
    names.append(products_key['name'].values[0])
    ratings.append(products_key['rating'].values[0])
    rating_counts.append(products_key['rating_count'].values[0])

data['author'] = author
data['brand'] = brand
data['name'] = names
data['rating'] = ratings
data['rating_count'] = rating_counts
data['stars'] = stars
data['title'] = title
data['text'] = text

Now we take a peek at the data and it's shape.

data.head(3)

data.shape

## Inspecting Nulls And Data Types

Now that there is one complete dataset, I can start checking out some of the data. I'll start with looking at some of the data types.

data.dtypes

All these data types seem okay to me. Now, I can check the null values.

for column in data:
    print(column + ": ", sum(data[column].isnull()))

So there are quite a few nulls in `title` and in `author`. That is okay. I wasn't really planning on using `author` anyways, I just have it for reference. As for `title`, that is alright too. The key data lives in `text`, which, luckily, doesn't contain any nulls. To deal with the nulls, I'll just fill them with empty strings.

data['title'].fillna('', inplace = True)

## A Little Text Preprocessing

I have two main columns with substantial textual data, `title` and `text`. For EDA and modelling, it is best to combine these into a single column. After that is done, I can remove stopwords, make everything lowercase and remove any special characers.

data['text'] = data[['title', 'text']].apply(' '.join, axis=1)

data.drop(['title'], axis=1, inplace=True)

data.head(3)

I will also clean the text up a bit and remove the stop words, so I can have a more usable text column.

def cleanText(text):
    text = re.sub(r'https?:\/\/\S+', '', text) # remove links
    text = re.sub(r'@[A-Za-z0-0]+', '', text) # remove @
    text = re.sub(r'[!@#$%^&*()_\-+=}{[\]|,.?<>:;\'’`~]', '', text) # remove special characters
    text = re.sub(r'[0-9]+[a-zA-Z]*', '', text) # remove numbers
    return text

data['text'] = data['text'].apply(cleanText)

data['text'] = data['text'].apply(lambda words: ' '.join(word.lower() for word in words.split() if word not in stop))

I'll also use a tokenizer and get the lemma from each word. This should make the modelling and sentiment analysis a bit easier.

tokenizer = WhitespaceTokenizer()
lemmatizer = WordNetLemmatizer()

def lemmatize_text(text):
    return [lemmatizer.lemmatize(w) for w in tokenizer.tokenize(text)]

data['text'] = data['text'].apply(lemmatize_text)

data['text'] = [' '.join(map(str, l)) for l in data['text']]

data.head(3)

One last step. I am going to do a bit of feature engineering here. 

Ratings are given stars on a scale from 1 to 5. This would therefore give 5 different potential classes for each record. This makes this problem into a multi classification problem. I am not sure yet if I want to do a multi classification problem or a binary classification problem yet, so I am going to create a new column, `good_review`, that will take the value `Good` if `stars` is 4 or 5, and `Bad` otherwise

data['good_review'] = np.where((data.stars == 4) | (data.stars==5), "Good", "Bad")

data.head(4)

Great! This looks like a pretty solid data set to kick things off with. I'll save this as a csv and move on to the EDA.

# Save Data

save_file(data, 'ice_cream_data.csv', '../data')

sk_prefix = "sagemaker/ice-cream-reviews/sklearncontainer"
ice_cream_data_path = sess.upload_data(
    path="../data/ice_cream_data.csv", bucket=bucket, key_prefix=sk_prefix
)
