==> notebooks/5_sagemaker.ipynb <==
from sagemaker.sklearn.estimator import SKLearn
import sagemaker
import boto3

FRAMEWORK_VERSION = "0.23-1"


# Configure the SKLearn estimator
sklearn_estimator = SKLearn(
    entry_point="script.py",
    role="arn:aws:iam::187704205293:role/sagemaker",
    instance_count=1,
    instance_type="ml.m5.large",
    framework_version=FRAMEWORK_VERSION,
    base_job_name="LR-custom-sklearn",
    hyperparameters={
        "l1_ratio": 0,
    },
    use_spot_instances=True,
    max_wait=7200,
    max_run=3600
)

# Create a SageMaker session
sess = sagemaker.Session()

# Specify the S3 bucket and prefix
bucket = 'ice-cream-reviews-sagemakerbucket'
print('Using bucket ' + bucket)

sk_prefix = "sagemaker/ice-cream-reviews-sagemakerbucket/sklearncontainer"
trainpath = "s3://ice-cream-reviews-sagemakerbucket/sagemaker/ice-cream-reviews/sklearncontainer/train.csv"
testpath = "s3://ice-cream-reviews-sagemakerbucket/sagemaker/ice-cream-reviews/sklearncontainer/test.csv"

# Launch the training job with asynchronous call
sklearn_estimator.fit({"train": trainpath, "test": testpath}, wait=True)

sklearn_estimator.latest_training_job.wait(logs="None")
sm_boto3 = boto3.client("sagemaker")
artifact = sm_boto3.describe_training_job(
    TrainingJobName=sklearn_estimator.latest_training_job.name
)["ModelArtifacts"]["S3ModelArtifacts"]

print("Model artifact persisted at " + artifact)

artifact

from sagemaker.sklearn.model import SKLearnModel
from time import gmtime, strftime

model_name = "Custom-sklearn-model-" + strftime("%Y-%m-%d-%H-%M-%S", gmtime())
model = SKLearnModel(
    name =  model_name,
    model_data=artifact,
    role="arn:aws:iam::187704205293:role/sagemaker",
    entry_point="script.py",
    framework_version=FRAMEWORK_VERSION,
)

model_name

# Endpoint deployment
endpoint_name = "Custom-sklearn-model-" + strftime("%Y-%m-%d-%H-%M-%S", gmtime())
print("EndpointName={}".format(endpoint_name))

predictor = model.deploy(
    initial_instance_count=1,
    instance_type="ml.m4.xlarge",
    endpoint_name=endpoint_name,
)

endpoint_name

sm_boto3.delete_endpoint(EndpointName=endpoint_name)
